updates details for brave search (#99)
* updates details for brave search

* resolves server-tool enumeration

Docker's mcp-registry/internal/mcp/client.go file pulls values from the env.example key when standing up the server, prior to requesting the tool list. The mcp-registry expects stdio to be the transport whereon tools are listed, but the "http" example in server.yaml causes the MCP Server to launch with an HTTP Transport, thus breaking the flow.