ci: add clean task and update validation script to use it
When modifying large numbers of server definitions (e.g. in #328), the
validation step can cause the worker to run out of disk space. This PR
adds a new clean command and updates the validation step to use it in
every call to process_server. This should allow a theoretically
unlimited number of validations to occur.

Signed-off-by: Jacob Howard <jacob.howard@docker.com>