ci: modify CI validation workflow to run tooling from main
As our server definition requirements evolve, we want to make sure
whatever is going into main is up-to-date with the latest requirements.
In order to accomplish that, we can't run the validation tooling from
the PR branch - instead we need to build it from main and then run it on
the PR worktree.

Signed-off-by: Jacob Howard <jacob.howard@docker.com>